--- testing/.gitignore
@@ -1,4 +1,0 @@
-*.so
-/vendor
-/bin
-/tmp
