--- testing/Gopkg.toml
@@ -1,18 +1,0 @@
-required = [
-  "github.com/golang/protobuf/protoc-gen-go",
-  "github.com/grpc-ecosystem/grpc-gateway/protoc-gen-grpc-gateway",
-  "github.com/grpc-ecosystem/grpc-gateway/protoc-gen-swagger",
-]
-
-[prune]
-  go-tests = true
-  unused-packages = true
-
-  [[prune.project]]
-    name = "github.com/grpc-ecosystem/grpc-gateway"
-    non-go = false
-    unused-packages = false
-
-[[constraint]]
-  branch = "master"
-  name = "github.com/izumin5210/grapi"
