--- testing/app/run.go
@@ -1,16 +1,0 @@
-package app
-
-import (
-	"github.com/izumin5210/grapi/pkg/grapiserver"
-)
-
-// Run starts the grapiserver.
-func Run() error {
-	s := grapiserver.New(
-		grapiserver.WithDefaultLogger(),
-		grapiserver.WithServers(
-		// TODO
-		),
-	)
-	return s.Serve()
-}
