--- testing/cmd/server/run.go
@@ -1,19 +1,0 @@
-package main
-
-import (
-	"os"
-
-	"github.com/izumin5210/grapi/testing/app"
-)
-
-func main() {
-	os.Exit(run())
-}
-
-func run() int {
-	err := app.Run()
-	if err != nil {
-		return 1
-	}
-	return 0
-}
