--- testing/grapi.toml
@@ -1,25 +1,0 @@
-[grapi]
-server_dir = "./app/server"
-
-[protoc]
-protos_dir = "./api/protos"
-out_dir = "./api"
-import_dirs = [
-  "./vendor/github.com/grpc-ecosystem/grpc-gateway",
-  "./vendor/github.com/grpc-ecosystem/grpc-gateway/third_party/googleapis",
-]
-
-  [[protoc.plugins]]
-  path = "./vendor/github.com/golang/protobuf/protoc-gen-go"
-  name = "go"
-  args = { plugins = "grpc" }
-
-  [[protoc.plugins]]
-  path = "./vendor/github.com/grpc-ecosystem/grpc-gateway/protoc-gen-grpc-gateway"
-  name = "grpc-gateway"
-  args = { logtostderr = true }
-
-  [[protoc.plugins]]
-  path = "./vendor/github.com/grpc-ecosystem/grpc-gateway/protoc-gen-swagger"
-  name = "swagger"
-  args = { logtostderr = true }
